Moved Logger to core
