+++ src/core/index.ts
@@ -1,10 +1,13 @@
+import { Logger } from "./logger";
 import { NoteGraph } from "./NoteGraph";
-import { MarkdownLink } from "./Link";
+import { Link, MarkdownLink } from "./Link";
 import { Note} from "./Note";
 import { CodekastenParser } from "./Parser";
 
 export { 
     CodekastenParser, 
+    Link,
+    Logger, 
     MarkdownLink,
     Note,
     NoteGraph
