+++ src/services/index.ts
@@ -1,4 +1,3 @@
-import { CodekastenLogger, Logger } from './logger';
 import { FilesystemSyncher } from './filesystemSyncher';
 import { bootstrap} from './bootstrap';
 import { config } from './config';
@@ -6,7 +5,5 @@ import { config } from './config';
 export {
     bootstrap,
     config,
-    CodekastenLogger, 
-    FilesystemSyncher,
-    Logger
+    FilesystemSyncher
 };
