+++ src/vscode/index.ts
@@ -1,0 +1,5 @@
+import { VscodeLogger } from "./VscodeLogger";
+
+export {
+    VscodeLogger
+};
